refactor: transform call-based services to entity-based services
